Add suggestions from code review

+++ src/main/java/seedu/duke/component/gates/Gate.java
@@ -6,7 +6,7 @@ public abstract class Gate {
 
     public abstract int getOutput();
 
-    public void setInput(byte input) {
+    public void setInput(int input) {
         this.input = input;
     }
